requirements.txt
aiogram>=3.13.0
aiohttp
telethon
google-generativeai